+++ components/transition/Modal.vue
@@ -1,0 +1,12 @@
+<template>
+  <transition
+    enter-active-class="ease-out"
+    enter-class="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
+    enter-to-class="opacity-100 translate-y-0 sm:scale-100"
+    leave-active-class="ease-in"
+    leave-class="opacity-100 translate-y-0 sm:scale-100"
+    leave-to-class="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
+  >
+    <slot></slot>
+  </transition>
+</template>
